Jogosultság ellenőrzés
Ads:
    - createAd
    - updateAd
    - deleteAd
    - getAllAds
    - getAllNonAcceptedAds
    - acceptAd
    - getAllAdsByUserId
AdsCounties:
    - addNewCountyToAd
    - deleteCountyFromAd
Jobs:
    - deleteJob
    - updateJobByWorker
    - updateJobByCustomer
    - getAllJobsByCustomer
    - getAllJobsByWorker
    - getAllJobs
    - acceptByWorker
    - acceptByCustomer
    - changeJobStatus
    - getJobById

+++ Backend/SzakemberKereso/src/main/java/szakemberkereso/Controller/AdsController.java
@@ -4,7 +4,9 @@
  */
 package szakemberkereso.Controller;
 
+import java.util.ArrayList;
 import java.util.List;
+import java.util.Objects;
 import javax.ws.rs.core.Context;
 import javax.ws.rs.core.UriInfo;
 import javax.ws.rs.Produces;
@@ -16,8 +18,11 @@ import javax.ws.rs.PUT;
 import javax.ws.rs.PathParam;
 import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
+import szakemberkereso.Configuration.Roles;
 import szakemberkereso.Model.Ads;
+import szakemberkereso.Model.Users;
 import szakemberkereso.Service.AdsService;
+import szakemberkereso.Service.AuthService;
 
 /**
  * REST Web Service
@@ -62,16 +67,31 @@ public class AdsController {
     @Path("createAd")
     @Consumes(MediaType.APPLICATION_JSON)
     public Response createAd(Ads ad){
-        Integer result = as.createAd(ad);
-        return Response.status(Response.Status.OK).entity(result).type(MediaType.APPLICATION_JSON).build();
+        if (AuthService.isUserAuthorized(ad.getCurrentUserId(), new Roles[]{Roles.ADMIN, Roles.WORKER})) {
+            //a USER jogosultságú user-ek nem hozhatnak létre hirdetéseket
+            Integer result = as.createAd(ad);
+            return Response.status(Response.Status.OK).entity(result).type(MediaType.APPLICATION_JSON).build();
+        }
+        else{
+            return Response.status(Response.Status.FORBIDDEN).entity("Nincs jogosultsága ehhez a kéréshez.").build();
+        }
     }
     
     @POST
     @Path("updateAd")
     @Consumes(MediaType.APPLICATION_JSON)
     public Response updateAd(Ads ad){
-        String result = as.updateAd(ad);
-        return Response.status(Response.Status.OK).entity(result).type(MediaType.APPLICATION_JSON).build();
+        if (AuthService.isUserAuthorized(ad.getCurrentUserId(), new Roles[]{Roles.ADMIN, Roles.WORKER})) {
+            //a user-ek csak a saját hirdetésüket módosíthatják
+            if(!Objects.equals(ad.getCurrentUserId(), Ads.getAdsById(ad.getId()).getUserId())){
+                return Response.status(Response.Status.FORBIDDEN).entity("Nincs jogosultsága ehhez a kéréshez.").build();
+            }
+            String result = as.updateAd(ad);
+            return Response.status(Response.Status.OK).entity(result).type(MediaType.APPLICATION_JSON).build();
+        }
+        else{
+            return Response.status(Response.Status.FORBIDDEN).entity("Nincs jogosultsága ehhez a kéréshez.").build();
+        }
     }
     
     @GET
@@ -85,32 +105,63 @@ public class AdsController {
     @Path("getAllNonAcceptedAds")
     @Consumes(MediaType.APPLICATION_JSON)
     public Response getAllNonAcceptedAds(Integer userId){
-        List<Ads> result = as.getAllNonAcceptedAds();
-        return Response.status(Response.Status.OK).entity(result).type(MediaType.APPLICATION_JSON).build();
+        //a nem ADMIN jogosultságú user-ek nem kérhetik le
+        if (AuthService.isUserAuthorized(userId, new Roles[]{Roles.ADMIN})) {
+            List<Ads> result = as.getAllNonAcceptedAds();
+            return Response.status(Response.Status.OK).entity(result).type(MediaType.APPLICATION_JSON).build();
+        }
+        else{
+            return Response.status(Response.Status.FORBIDDEN).entity("Nincs jogosultsága ehhez a kéréshez.").build();
+        }
     }
     
     @POST
     @Path("acceptAd")
     @Consumes(MediaType.APPLICATION_JSON)
     public Response acceptAd(Ads ad){
-        Boolean result = as.acceptAd(ad.getId());
-        return Response.status(Response.Status.OK).entity(result).type(MediaType.APPLICATION_JSON).build();
+        //a nem ADMIN jogosultságú user-ek nem fogadhatják el
+        if (AuthService.isUserAuthorized(ad.getCurrentUserId(), new Roles[]{Roles.ADMIN})) {
+            Boolean result = as.acceptAd(ad.getId());
+            return Response.status(Response.Status.OK).entity(result).type(MediaType.APPLICATION_JSON).build();
+        }
+        else{
+            return Response.status(Response.Status.FORBIDDEN).entity("Nincs jogosultsága ehhez a kéréshez.").build();
+        }
     }
     
     @POST
     @Path("getAllAds")
     @Consumes(MediaType.APPLICATION_JSON)
     public Response getAllAds(Integer userId){
-        List<Ads> result = as.getAllAds();
-        return Response.status(Response.Status.OK).entity(result).type(MediaType.APPLICATION_JSON).build();
+        if (AuthService.isUserAuthorized(userId, new Roles[]{Roles.ADMIN})){
+            List<Ads> result = as.getAllAds();
+            return Response.status(Response.Status.OK).entity(result).type(MediaType.APPLICATION_JSON).build();
+        }
+        else{
+            return Response.status(Response.Status.FORBIDDEN).entity("Nincs jogosultsága ehhez a kéréshez.").build();
+        }
     }
     
     @POST
     @Path("getAllAdsByUserId")
     @Consumes(MediaType.APPLICATION_JSON)
     public Response getAllAdsByUserId(Ads ad){
-        List<Ads> result = as.getAllAdsByUserId(ad.getUserId());
-        return Response.status(Response.Status.OK).entity(result).type(MediaType.APPLICATION_JSON).build();
+        if (AuthService.isUserAuthorized(ad.getCurrentUserId(), new Roles[]{Roles.ADMIN, Roles.WORKER})) {
+            List<Ads> result = as.getAllAdsByUserId(ad.getUserId());
+            return Response.status(Response.Status.OK).entity(result).type(MediaType.APPLICATION_JSON).build();
+        }
+        else if(AuthService.isUserAuthorized(ad.getCurrentUserId(), new Roles[]{Roles.USER})){
+            //a USER jogosultságú user-ek csak az elfogadott hirdetéseket láthatják
+            List<Ads> ads = as.getAllAdsByUserId(ad.getUserId());
+            List<Ads> result = new ArrayList<>();
+            for(Ads a : ads){
+                if(a.getStatus() == 1){
+                    result.add(a);
+                }
+            }
+            return Response.status(Response.Status.OK).entity(result).type(MediaType.APPLICATION_JSON).build();
+        }
+        return Response.status(Response.Status.UNAUTHORIZED).entity("Nem sikerült azonosítani.").build();
     }
     
     @POST
@@ -132,8 +183,17 @@ public class AdsController {
     @Path("deleteAd")
     @Consumes(MediaType.APPLICATION_JSON)
     public Response deleteAd(Ads ad){
-        Boolean result = as.deleteAd(ad.getId());
-        return Response.status(Response.Status.OK).entity(result).type(MediaType.APPLICATION_JSON).build();
+        if (AuthService.isUserAuthorized(ad.getCurrentUserId(), new Roles[]{Roles.ADMIN, Roles.USER, Roles.WORKER})) {
+            if(!AuthService.isUserAuthorized(ad.getCurrentUserId(), new Roles[]{Roles.ADMIN})){
+                //a nem ADMIN jogosultságú user-ek csak a saját hirdetésüket törölhetik
+                if(!Objects.equals(ad.getCurrentUserId(), Ads.getAdsById(ad.getId()).getUserId())){
+                    return Response.status(Response.Status.FORBIDDEN).entity("Nincs jogosultsága ehhez a kéréshez.").build();
+                }
+            }
+            Boolean result = as.deleteAd(ad.getId());
+            return Response.status(Response.Status.OK).entity(result).type(MediaType.APPLICATION_JSON).build();
+        }
+        return Response.status(Response.Status.UNAUTHORIZED).entity("Nem sikerült azonosítani.").build();
     }
     
 }
