+++ Backend/SzakemberKereso/src/main/java/szakemberkereso/Controller/AdsCountiesController.java
@@ -5,6 +5,7 @@
 package szakemberkereso.Controller;
 
 import java.util.List;
+import java.util.Objects;
 import javax.ws.rs.core.Context;
 import javax.ws.rs.core.UriInfo;
 import javax.ws.rs.Produces;
@@ -16,9 +17,12 @@ import javax.ws.rs.PUT;
 import javax.ws.rs.PathParam;
 import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
+import szakemberkereso.Configuration.Roles;
+import szakemberkereso.Model.Ads;
 import szakemberkereso.Model.AdsCounties;
 import szakemberkereso.Model.Counties;
 import szakemberkereso.Service.AdsCountiesService;
+import szakemberkereso.Service.AuthService;
 
 /**
  * REST Web Service
@@ -63,16 +67,31 @@ public class AdsCountiesController {
     @Path("addNewCountyToAd")
     @Consumes(MediaType.APPLICATION_JSON)
     public Response addNewCountyToAd(AdsCounties ad_county){
-        String result = acs.addNewCountyToAd(ad_county);
-        return Response.status(Response.Status.OK).entity(result).type(MediaType.APPLICATION_JSON).build();
+        if (AuthService.isUserAuthorized(ad_county.getCurrentUserId(), new Roles[]{Roles.WORKER})) {
+            //csak WORKER jogosultságú user-ek és csak a saját hirdetésükhöz adhatnak hozzá megyét
+            if(!Objects.equals(ad_county.getCurrentUserId(), Ads.getAdsById(ad_county.getAdId()).getUserId())){
+                return Response.status(Response.Status.FORBIDDEN).entity("Nincs jogosultsága ehhez a kéréshez.").build();
+            }
+            String result = acs.addNewCountyToAd(ad_county);
+            return Response.status(Response.Status.OK).entity(result).type(MediaType.APPLICATION_JSON).build();
+        }
+        return Response.status(Response.Status.UNAUTHORIZED).entity("Nem sikerült azonosítani.").build();
     }
     
     @POST
     @Path("deleteCountyFromAd")
     @Consumes(MediaType.APPLICATION_JSON)
     public Response deleteCountyFromAd(AdsCounties ad_county){
-        String result = acs.deleteCountyFromAd(ad_county);
-        return Response.status(Response.Status.OK).entity(result).type(MediaType.APPLICATION_JSON).build();
+        if (AuthService.isUserAuthorized(ad_county.getCurrentUserId(), new Roles[]{Roles.WORKER})) {
+            //csak WORKER jogosultságú user-ek és csak a saját hirdetésükből törölhetnek megyét
+            if(!Objects.equals(ad_county.getCurrentUserId(), Ads.getAdsById(ad_county.getAdId()).getUserId())){
+                return Response.status(Response.Status.FORBIDDEN).entity("Nincs jogosultsága ehhez a kéréshez.").build();
+            }
+            String result = acs.deleteCountyFromAd(ad_county);
+            return Response.status(Response.Status.OK).entity(result).type(MediaType.APPLICATION_JSON).build();
+        }
+        return Response.status(Response.Status.UNAUTHORIZED).entity("Nem sikerült azonosítani.").build();
+        
     }
     
     @GET
